int type for position
