+++ pynpoint/processing/extract.py
@@ -12,8 +12,7 @@ import numpy as np
 from typeguard import typechecked
 
 from pynpoint.core.processing import ProcessingModule
-from pynpoint.util.module import locate_star, rotate_coordinates
-from pynpoint.util.image import crop_image, center_pixel
+from pynpoint.util.image import crop_image, center_pixel, locate_star, rotate_coordinates
 
 
 class StarExtractionModule(ProcessingModule):
