+++ pynpoint/processing/pcabackground.py
@@ -19,7 +19,8 @@ from pynpoint.core.processing import ProcessingModule
 from pynpoint.processing.resizing import CropImagesModule
 from pynpoint.processing.stacksubset import CombineTagsModule
 from pynpoint.processing.psfpreparation import SortParangModule
-from pynpoint.util.module import progress, memory_frames, locate_star
+from pynpoint.util.image import locate_star
+from pynpoint.util.module import progress, memory_frames
 
 
 class PCABackgroundPreparationModule(ProcessingModule):
