+++ tests/test_processing/test_fluxposition.py
@@ -230,7 +230,7 @@ class TestFluxPosition:
                                            psf_in_tag='read',
                                            res_out_tag='simplex_res',
                                            flux_position_tag='flux_position',
-                                           position=(31., 49.),
+                                           position=(31, 49),
                                            magnitude=6.,
                                            psf_scaling=-1.,
                                            merit='hessian',
@@ -248,17 +248,17 @@ class TestFluxPosition:
 
         data = self.pipeline.get_data('simplex_res')
         assert np.allclose(data[0, 50, 31], 0.00012976212788352575, rtol=limit, atol=0.)
-        assert np.allclose(data[42, 50, 31], 1.2141761821389107e-05, rtol=limit, atol=0.)
-        assert np.allclose(np.mean(data), 9.461337432531517e-09, rtol=limit, atol=0.)
-        assert data.shape == (43, 101, 101)
+        assert np.allclose(data[41, 50, 31], 1.7176632474300478e-05, rtol=limit, atol=0.)
+        assert np.allclose(np.mean(data), 9.470306099078656e-09, rtol=limit, atol=0.)
+        assert data.shape == (42, 101, 101)
 
         data = self.pipeline.get_data('flux_position')
-        assert np.allclose(data[42, 0], 31.6456737445356, rtol=limit, atol=0.)
-        assert np.allclose(data[42, 1], 49.9199601480223, rtol=limit, atol=0.)
-        assert np.allclose(data[42, 2], 0.49557152090327206, rtol=limit, atol=0.)
-        assert np.allclose(data[42, 3], 90.24985480686087, rtol=limit, atol=0.)
-        assert np.allclose(data[42, 4], 5.683191873535635, rtol=limit, atol=0.)
-        assert data.shape == (43, 6)
+        assert np.allclose(data[41, 0], 31.67747653034013, rtol=limit, atol=0.)
+        assert np.allclose(data[41, 1], 49.91874032275719, rtol=limit, atol=0.)
+        assert np.allclose(data[41, 2], 0.4947129988341501, rtol=limit, atol=0.)
+        assert np.allclose(data[41, 3], 90.25410288237038, rtol=limit, atol=0.)
+        assert np.allclose(data[41, 4], 5.691970059845411, rtol=limit, atol=0.)
+        assert data.shape == (42, 6)
 
     def test_simplex_minimization_reference(self):
 
@@ -267,7 +267,7 @@ class TestFluxPosition:
                                            psf_in_tag='read',
                                            res_out_tag='simplex_res_ref',
                                            flux_position_tag='flux_position_ref',
-                                           position=(31., 49.),
+                                           position=(31, 49),
                                            magnitude=6.,
                                            psf_scaling=-1.,
                                            merit='sum',
@@ -285,17 +285,17 @@ class TestFluxPosition:
 
         data = self.pipeline.get_data('simplex_res_ref')
         assert np.allclose(data[0, 50, 31], 0.00014188043631450017, rtol=limit, atol=0.)
-        assert np.allclose(data[35, 50, 31], 8.260705332688204e-05, rtol=limit, atol=0.)
-        assert np.allclose(np.mean(data), 1.3084672332295342e-06, rtol=limit, atol=0.)
-        assert data.shape == (36, 101, 101)
+        assert np.allclose(data[35, 50, 31], 6.660723685443821e-05, rtol=limit, atol=0.)
+        assert np.allclose(np.mean(data), 1.3228426986034557e-06, rtol=limit, atol=0.)
+        assert data.shape == (44, 101, 101)
 
         data = self.pipeline.get_data('flux_position_ref')
-        assert np.allclose(data[35, 0], 31.523599108957953, rtol=limit, atol=0.)
-        assert np.allclose(data[35, 1], 49.80956476439131, rtol=limit, atol=0.)
-        assert np.allclose(data[35, 2], 0.49888932122698404, rtol=limit, atol=0.)
-        assert np.allclose(data[35, 3], 90.59052349996864, rtol=limit, atol=0.)
-        assert np.allclose(data[35, 4], 6.021039673141669, rtol=limit, atol=0.)
-        assert data.shape == (36, 6)
+        assert np.allclose(data[43, 0], 31.519027018276986, rtol=limit, atol=0.)
+        assert np.allclose(data[43, 1], 49.85541939005469, rtol=limit, atol=0.)
+        assert np.allclose(data[43, 2], 0.4990015399214498, rtol=limit, atol=0.)
+        assert np.allclose(data[43, 3], 90.44822801080721, rtol=limit, atol=0.)
+        assert np.allclose(data[43, 4], 5.9953709695084605, rtol=limit, atol=0.)
+        assert data.shape == (44, 6)
 
     def test_mcmc_sampling_gaussian(self):
 
@@ -352,8 +352,8 @@ class TestFluxPosition:
                                     psf_scaling=-1.,
                                     pca_number=1,
                                     aperture={'type':'circular',
-                                              'pos_x':7.0,
-                                              'pos_y':12.5,
+                                              'pos_x':7,
+                                              'pos_y':12,
                                               'radius':0.1},
                                     mask=None,
                                     extra_rot=0.,
@@ -392,12 +392,10 @@ class TestFluxPosition:
                                     nsteps=150,
                                     psf_scaling=-1.,
                                     pca_number=1,
-                                    aperture={'type':'elliptical',
-                                              'pos_x':7.0,
-                                              'pos_y':12.5,
-                                              'semimajor':0.1,
-                                              'semiminor':0.1,
-                                              'angle':0.0},
+                                    aperture={'type':'circular',
+                                              'pos_x':7,
+                                              'pos_y':12,
+                                              'radius':0.1},
                                     mask=None,
                                     extra_rot=0.,
                                     scale=2.,
@@ -436,8 +434,8 @@ class TestFluxPosition:
                                     psf_scaling=-1.,
                                     pca_number=1,
                                     aperture={'type':'circular',
-                                              'pos_x':7.0,
-                                              'pos_y':12.5,
+                                              'pos_x':7,
+                                              'pos_y':12,
                                               'radius':0.1},
                                     mask=None,
                                     extra_rot=0.,
